api/requirements.txt
fastapi==0.78.0
mongoengine==0.24.1
python-dotenv==0.20.0
python_dateutil==2.8.2
starlette==0.19.1
uvicorn==0.17.6
dnspython==2.2.1
